% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{ari}
\alias{ari}
\title{Adjusted Rand Index}
\usage{
ari(labels.ref, labels.est)
}
\arguments{
\item{labels.ref}{Reference label vector}

\item{labels.est}{Estimated label vector}
}
\value{
calculated ARI
}
\description{
Adjusted Rand Index
}
