% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{calculate_alignment}
\alias{calculate_alignment}
\title{Calculate the alginment score}
\usage{
calculate_alignment(
  X.list,
  k = NULL,
  balanced = F,
  min.sample.per.dataset = 100,
  dataset = NULL,
  labels = NULL,
  max.k = 50,
  seed = 0,
  verbose = F,
  pca = F
)
}
\arguments{
\item{X.list}{a list of gene expression matrix per batch}

\item{k}{integer, number of nearest neighbor to consider}

\item{balanced}{boolean, whether to subset to balance the size of baches}

\item{min.sample.per.dataset}{integer, minimum cells to sample per dataset (default 100)}

\item{dataset}{factor/characteristic vector that indicate the batch of cells if len(X.list)==1}

\item{labels}{factor/characteristic vector, cluster/cell type labels of the cells}

\item{max.k}{integer, maximum number of nearest neighbors to consider (cap for large dataset)}

\item{seed}{random seed}

\item{verbose}{boolean scalar, whether to show extensive program logs (default TRUE)}

\item{pca}{if pca!=F, perform pca first before calculating the pairwise distance}
}
\value{
A list containing \describe{
\item{alignment.per.sample} a numeric vector, alignment score per cell
\item{alignment.per.cluster} a numeric vector, alignment score per cluster. Only calculated
with labels are provided
\item{alignment.per.dataset} a numeric vector, alignment score per dataset.
}
}
\description{
Calculate the alginment score per sample, per cluster and per dataset
}
