% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{calculate_alignment_per_type}
\alias{calculate_alignment_per_type}
\title{Calculate the alginment score}
\usage{
calculate_alignment_per_type(
  X,
  dataset = NULL,
  labels = NULL,
  k = NULL,
  max.k = 50,
  seed = 0,
  verbose = F,
  pca = F
)
}
\arguments{
\item{X}{ncell-by-ngene expression matrix}

\item{dataset}{factor/characteristic vector that indicate the batch of cells}

\item{labels}{factor/characteristic vector, cluster/cell type labels of the cells}

\item{k}{integer, number of nearest neighbor to consider}

\item{max.k}{integer, maximum number of nearest neighbors to consider (cap for large dataset)}

\item{seed}{random seed}

\item{verbose}{boolean scalar, whether to show extensive program logs (default TRUE)}

\item{pca}{if pca!=F, perform pca first before calculating the pairwise distance}
}
\value{
A list containing \describe{
\item{alignment.per.sample}{a numeric vector, alignment score per cell}
\item{alignment.per.cluster}{a numeric vector, alignment score per cluster. Only calculated
with labels are provided}
\item{alignment.per.dataset}{a numeric vector, alignment score per dataset.}
\item{alignment.per.dataset.cluster}{numeric vector, alignment score per dataset per cluster}
}
}
\description{
Calculate the alginment score per cluster, where the normalization is performed based on the
per-group batch population ratios
}
