% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transfer.R
\name{CFITTransfer}
\alias{CFITTransfer}
\title{Transfer the learned gene expression signature onto a new dataset}
\usage{
CFITTransfer(
  Xtarget,
  Wref,
  max.niter = 100,
  tol = 1e-05,
  init = NULL,
  seed = 0,
  verbose = T,
  n.cores = parallel::detectCores() - 4
)
}
\arguments{
\item{Xtarget}{An ncells-by-ngenes gene expression matrix}

\item{Wref}{An ngenes-by-r reference low dimensional factor matrix}

\item{max.niter}{integer, max number of iterations (default 100).}

\item{tol}{numeric scalar, tolerance used in stopping criteria (default 1e-5).}

\item{init}{a list of parameters for parameter initialization: lambda, b, H}

\item{seed}{random seed used (default 0)}

\item{verbose}{boolean scalar, whether to show extensive program logs (default TRUE)}

\item{n.cores}{number of cores used for parallel computing}
}
\value{
a list containing  \describe{
\item{H}{estimated factor loading matrix of size ncells-by-r}
\item{b}{estimated shift vector of size p (ngenes)}
\item{lambda}{estimated scaling vector of size p (ngenes)}
\item{convergence}{boolean, whether the algorithm converge}
\item{obj}{numeric scalar, value of the objective function at convergence or when maximum iteration achieved}
\item{niter}{integer, the iteration at convergence (or maximum iteration if not converge)}
\item{delta}{numeric scalar, the relative difference in last objective update}
\item{params}{list of parameters used for the algorithm: max.iter, tol, nrep}
}
}
\description{
Estimate the data specific parameters through Iterative Nonnegative Matrix Factorization (iNMF),
by minimizing the objective function \deqn{1/n||X -(HW^T\Lambda  + 1_n b^T)||_F^2}.
}
