% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_by_model.R
\name{generate_centers}
\alias{generate_centers}
\title{Generate gaussian centers with fixed minimum separation c}
\usage{
generate_centers(K, d, sig, cl.sep, max.m = 1000)
}
\arguments{
\item{K}{number of clusters}

\item{d}{dimension of space where K centers are generated}

\item{sig}{variance of within cluster variance that affects the cluster separation}

\item{cl.sep}{cluster separation}

\item{max.m}{max number of attemps}
}
\value{
A list containing \describe{
\item{centers} generated K-by-d corresponding to K centers in d-dim space
\item{status} boolean indicating whether the generationg is successful
}
}
\description{
Generate K centers in d dimenaional space such that min ||mu_i-mu_j|| = c\sqrt(d, sig)
}
