% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_by_model.R
\name{generate_data}
\alias{generate_data}
\title{Simulate data from the cFIT model}
\usage{
generate_data(
  n,
  ntask,
  K,
  p,
  alpha = NULL,
  sig = 1,
  cl.sep = 1,
  batch.effect.sig = 0.1
)
}
\arguments{
\item{n}{number of data point per dataset}

\item{ntask}{number of batches}

\item{K}{number of clusters}

\item{p}{number of genes}

\item{alpha}{parameter for Dirichilet distribution used to generate the labels (default 10,
representing equal cluster size. smaller alpha corresponds to more unbalanced types)}

\item{sig}{with cluster variance}

\item{cl.sep}{Cluster center separation, the higher the clusters are well separated}

\item{batch.effect.sig}{batch effect variance, higher the large batch effects are}
}
\value{
A list of generated data, \describe{
\item{X.list} a list of n-by-p expression matrix
\item{H.list} a list of n-by-K binary membership matrix
\item{lambda.list} a list of length-p vectors of per-dataset scaling
\item{b.list} a list of length-p vectors of per-dataset shift
\item{E.list} a list of noise matrix
\item{label.list} a list of length-n vectors of cluster labels
\item{W} p-by-K common factor matrix
}
}
\description{
Simulate \deqn{X_j = H_jW\Lambda_j + 1_{n_j}b_j +E_j, j = 1,\ldots, ntask}. The nonnegative matrix
is generated as the cluster centers given number of clusters K. H_j is generated as the binary
membership matrix, where the labels are generated from a Dirichlet distribution with parameter
alpha. Distortion lambda and shift b are generated from truncated normal distribution, Noise
matrix E is generate with each entry from iid normal distribution.
}
