% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{getmode}
\alias{getmode}
\title{Get the the mode in a categorical vector}
\usage{
getmode(x)
}
\arguments{
\item{x}{a categorical vector}
}
\value{
mode in the vector
}
\description{
Get the the mode in a categorical vector
}
