% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing_utils.R
\name{human_gene_to_mouse_genes}
\alias{human_gene_to_mouse_genes}
\title{Find the mapped mouse gene of each human gene}
\usage{
human_gene_to_mouse_genes(gene.list)
}
\arguments{
\item{gene.list}{a character vector of human gene symbols, indicating which human genes are to match}
}
\value{
a list of vectors, corresponding to the mapped mouse genes for each human gene
}
\description{
Find the mapped mouse gene of each human gene
}
