% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing_utils.R
\name{match_mouse_expr_to_human_genes}
\alias{match_mouse_expr_to_human_genes}
\title{Match the mouse gene expressions to the human gene expressions}
\usage{
match_mouse_expr_to_human_genes(mouse.exprs, human.genes.list)
}
\arguments{
\item{mouse.exprs}{ncell-by-ngene expression matrix of mouse}

\item{human.genes.list}{a character vector of human gene symbols, indicating which genes are to match}
}
\value{
ncell-by-ngene' of mapped human expression matrix
}
\description{
Match the expression of mouse genes to expression of human genes.
If the case of multi-match, the expressions are averaged
}
