% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing_utils.R
\name{mouse_gene_to_human_genes}
\alias{mouse_gene_to_human_genes}
\title{Find the mapped human gene of each mouse gene}
\usage{
mouse_gene_to_human_genes(gene.list)
}
\arguments{
\item{gene.list}{a character vector of mouse gene symbols, indicating which mouse genes are to match}
}
\value{
a list of vectors, corresponding to the mapped human genes for each mouse gene
}
\description{
Find the mapped human gene of each mouse gene
}
