% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration.R
\name{objective_func}
\alias{objective_func}
\title{Calculate the objective function}
\usage{
objective_func(X.list, W, lambda.list, H.list, b.list)
}
\arguments{
\item{X.list}{a list of ncells-by-ngenes gene expression matrix}

\item{W}{ngenes-by-r numeric matrix.}

\item{lambda.list}{A list of scaling vector of size p (ngenes).}

\item{H.list}{A list of factor loading matrix of size ncells-by-r}

\item{b.list}{A list of shift vector of size p (ngenes).}
}
\value{
numeric scalar, the value of the objective function
}
\description{
\deqn{1/N \sum_j||X_j -(H_JW^T\Lambda_j  + 1_{n_j} b_j^T)||_F^2}
}
