% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing_utils.R
\name{preprocess_for_integration}
\alias{preprocess_for_integration}
\title{Preprocess expression matrix as input for data integration}
\usage{
preprocess_for_integration(
  X.list,
  genes,
  scale.factor = 10^4,
  scale = T,
  center = F
)
}
\arguments{
\item{X.list}{A list of m ncells-by-ngenes, gene expression matrices from m data sets}

\item{genes}{character vector, a vector of gene names (selected highly variable genes)}

\item{scale.factor}{scalar, scaling factor for library size normalization (default 1e4)}

\item{scale}{boolean, whether to scale per gene expression (default TRUE)}

\item{center}{boolean, whether to center per gene expression (default FALSE)}
}
\value{
preprocessed expression list X.list
}
\description{
Preprocess, leave highly expressed genes, depth normalized, log transform, scale but not centered
}
