% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing_utils.R
\name{preprocess_for_transfer}
\alias{preprocess_for_transfer}
\title{Preprocess expression matrix as input for data transfer}
\usage{
preprocess_for_transfer(
  counts,
  Wref,
  scale.factor = 10000,
  scale = T,
  center = F,
  verbose = F
)
}
\arguments{
\item{counts}{ncells-by-ngenes raw count matrix from target data}

\item{Wref}{ngenes-by-r reference factor matrix}

\item{scale.factor}{scalar, scaling factor for library size normalization (default 1e4)}

\item{scale}{boolean, whether to scale per gene expression (default TRUE)}

\item{center}{boolean, whether to center per gene expression (default FALSE)}

\item{verbose}{boolean scalar, whether to show extensive program logs (default TRUE)}
}
\value{
A list containing \describe{
\item{exprs}{ncell-by-ngenes preprocessed expresssion matrix}
\item{Wref}{ngenes-by-r reference factor matrix, with genes filtered}
}
}
\description{
Subset to the shared gene sets between target data and reference factor matrix.
Depth normalized, log transform, scale but not centered
}
