% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration.R
\name{solve_b}
\alias{solve_b}
\title{Solve for dataset specific shift b}
\usage{
solve_b(X, W, H, lambd)
}
\arguments{
\item{X}{ncells-by-ngenes gene expression matrix}

\item{W}{ngenes-by-r non-negative common factor matrix}

\item{H}{ncells-by-r nonnegative factor loading matrix}

\item{lambd}{numeric scalar, scaling associated with the dataset}
}
\value{
b shift vector of size p (ngenes).
}
\description{
\deqn{argmin_{b_j>=0} ||X- HW^T diag(lambda_j) - 1_n b^T||_F^2 }
}
