% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration_sketched.R
\name{solve_b_penalized}
\alias{solve_b_penalized}
\title{Solve for dataset specific shift b with Stochastic Porximal Point method}
\usage{
solve_b_penalized(X, W, H, lambd, b.old, iter)
}
\arguments{
\item{X}{ncells-by-ngenes gene expression matrix}

\item{W}{ngenes-by-r non-negative common factor matrix}

\item{H}{ncells-by-r nonnegative factor loading matrix}

\item{lambd}{numeric scalar, scaling associated with the dataset}

\item{b.old}{b vector  from last iteration}

\item{iter}{current iteration, for calculating the step size.}
}
\value{
b shift vector of size p (ngenes).
}
\description{
Solve for dataset specific shift b with Stochastic Porximal Point method
}
