% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration.R
\name{solve_H}
\alias{solve_H}
\title{Solve for factor loading paramters H}
\usage{
solve_H(X, W, lambd, b, n.cores)
}
\arguments{
\item{X}{ncells-by-ngenes gene expression matrix}

\item{W}{ngenes-by-r non-negative common factor matrix}

\item{lambd}{nonnegative numeric scalar, scaling associated with the dataset}

\item{b}{nonegative scalar, shift associated with tht edataset}
}
\value{
ncells-by-r matrix, factor loading matrix H
}
\description{
\deqn{argmin_{H>=0} ||X- HW^T diag(lambd) - 1_n b^T||_F^2   s.t H1 = 1)}
}
