% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration.R
\name{solve_lambda_list}
\alias{solve_lambda_list}
\title{Solve for dataset specific scalings lambda.list}
\usage{
solve_lambda_list(X.list, W, H.list, b.list, gamma, n.cores)
}
\arguments{
\item{X.list}{A list of ncells-by-ngenes gene expression matrix.}

\item{W}{ngenes-by-r non-negative common factor matrix}

\item{H.list}{A list of factor loading matrix of size ncells-by-r}

\item{b.list}{A list of shift vector of size p (ngenes).}

\item{gamma}{numeric scalar, parameter for the penalty term.}

\item{n.cores}{integer, number of cores used for parallel computation}
}
\value{
lambda.list A list of m scaling vector of size p (ngenes).
}
\description{
\deqn{argmin_{lambda_j>=0} ||X- HW^T diag(lambda_j) - 1_n b^T||_F^2 +
gamma * N * sum_{l=1}^p(\sum_{j=1}^mnj/N \lambda_{jl}-1)^2}
}
