% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration_speed_up.R
\name{solve_lambda_list_penalized}
\alias{solve_lambda_list_penalized}
\title{Solve for dataset specific scalings lambda.list with Stochastic Porximal Point method}
\usage{
solve_lambda_list_penalized(
  X.list,
  W,
  H.list,
  b.list,
  gamma,
  lambda.list.old,
  iter,
  n.cores
)
}
\arguments{
\item{X.list}{A list of ncells-by-ngenes gene expression matrix.}

\item{W}{ngenes-by-r non-negative common factor matrix}

\item{H.list}{A list of factor loading matrix of size ncells-by-r}

\item{b.list}{A list of shift vector of size p (ngenes).}

\item{gamma}{numeric scalar, parameter for the penalty term.}

\item{lambda.list.old}{lambda.list from last iteration}

\item{iter}{current iteration, for calculating the step size.}
}
\value{
lambda.list A list of m scaling vector of size p (ngenes).
}
\description{
Solve for dataset specific scalings lambda.list with Stochastic Porximal Point method
}
