% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration.R
\name{solve_subproblem}
\alias{solve_subproblem}
\title{Solve subproblems via coordinate descent, given the parameter to update}
\usage{
solve_subproblem(
  params.to.update = c("W", "lambda", "H"),
  X.list,
  W,
  H.list,
  b.list,
  lambda.list,
  verbose = T
)
}
\arguments{
\item{params.to.update}{a characteristic scalar, choice of ('W','lambda','b','H'),
specifying which set of parameters to update}

\item{X.list}{a list of ncells-by-ngenes gene expression matrix}

\item{W}{ngenes-by-r numeric matrix.}

\item{H.list}{A list of factor loading matrix of size ncells-by-r}

\item{b.list}{A list of shift vector of size p (ngenes).}

\item{lambda.list}{A list of scaling vector of size p (ngenes).}

\item{verbose}{boolean scalar, whether to show extensive program logs (default TRUE)}
}
\value{
a list containing updated parameters: W, H.list, lambda.list,  b.list
}
\description{
Solve the subproblem given which parameter set to update. For each subproblem,
the exact solution is obtained
}
