% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration_speed_up.R
\name{solve_subproblem_penalized}
\alias{solve_subproblem_penalized}
\title{Solve subproblems via coordinate descent with Stochastic proximal point method}
\usage{
solve_subproblem_penalized(
  params.to.update = c("W", "lambda", "b", "H"),
  X.list,
  W,
  H.list,
  b.list,
  lambda.list,
  gamma,
  iter,
  params.list.last,
  verbose = T,
  n.cores
)
}
\arguments{
\item{params.to.update}{a characteristic scalar, choice of ('W','lambda','b','H'),
specifying which set of parameters to update}

\item{X.list}{a list of ncells-by-ngenes gene expression matrix}

\item{W}{ngenes-by-r numeric matrix.}

\item{H.list}{A list of factor loading matrix of size ncells-by-r}

\item{b.list}{A list of shift vector of size p (ngenes).}

\item{lambda.list}{A list of scaling vector of size p (ngenes).}

\item{gamma}{numeric scalar, parameter for the penalty term.}

\item{iter}{integer, the current iteration}

\item{params.list.last}{The parameters from last iteration, for SPP}

\item{verbose}{boolean scalar, whether to show extensive program logs (default TRUE)}

\item{n.cores}{integer, the number of cores used for parallel computing}
}
\value{
a list containing updated parameters: W, H.list, lambda.list,  b.list
}
\description{
Solve the subproblem given which parameter set to update. For each subproblem,
the exact solution is obtained. The sketched objective function is solve w.r.t the set
of parameter with additional penalty in the form of \deqn{\frac{1}{\mu_t} \|w-w^{}t-1\|},
where the step size \deqn{mu_t = 0.1\times iter}.
}
