% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration.R
\name{solve_W}
\alias{solve_W}
\title{Solve for nonnegative common factor matrix W}
\usage{
solve_W(X.list, H.list, lambda.list, b.list, n.cores)
}
\arguments{
\item{X.list}{A list of ncells-by-ngenes gene expression matrix.}

\item{H.list}{A list of factor loading matrix of size ncells-by-r}

\item{lambda.list}{A list of scaling vector of size p (ngenes).}

\item{b.list}{A list of shift vector of size p (ngenes).}
}
\value{
W ngenes-by-r common factor matrix shared among datasets
}
\description{
\deqn{argmin_{W>=0} ||X- HW^T diag(lambd) - 1_n b^T||_F^2  }
}
