% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing_utils.R
\name{split_dataset_by_batch}
\alias{split_dataset_by_batch}
\title{Split the expression matrix into a list of per batch expressions}
\usage{
split_dataset_by_batch(
  X,
  batch,
  labels = NULL,
  metadata = NULL,
  dataset.name = NULL
)
}
\arguments{
\item{X}{An ncells-by-ngenes gene expression matrix}

\item{batch}{categorical/characteristic vector as batch id}

\item{labels}{(optional) charateristic vector of cell labels (default NULL)}

\item{metadata}{(optinal) dataframe of ncells row saving the cell metadata}

\item{dataset.name}{string, prefix of list name}
}
\value{
a list containing  \describe{
\item{X.list}{a list of gene expression matrix per batch}
\item{labels.list}{a list of vectors, saving the cell labels per batch}
\item{metadata.list}{a list of dataframe saving the metadata per batch}
}
}
\description{
Split the expression matrix into a list of per batch expressions
}
