% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration_sketched.R
\name{statistical_leverage_score}
\alias{statistical_leverage_score}
\title{Calculate the statistical leverage score}
\usage{
statistical_leverage_score(A, k = NULL)
}
\arguments{
\item{A}{data matrix}

\item{k}{number of singular vectors used for calculating scores}
}
\value{
a vector of scores per samples
}
\description{
Calculate the statistical leverage score
}
