% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration_sketched.R
\name{step_size}
\alias{step_size}
\title{Step size controller}
\usage{
step_size(iter, mu = 0.005)
}
\arguments{
\item{iter}{current iteration t}

\item{mu}{initial penalty when t=1}
}
\description{
Step size controller
}
