% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration_speed_up.R
\name{subsample}
\alias{subsample}
\title{Subsample from a list of data matrix}
\usage{
subsample(X.list, subsample.prop, min.samples = 20, weight.list = NULL)
}
\arguments{
\item{X.list}{a list of m ncells-by-ngenes, gene expression matrices from m data sets}

\item{subsample.prop}{a scalar between 0 and 1. smaller proportion with results in fast computation but less accurate results.}

\item{min.samples}{integer, minimum number of samples from each batch (20 by default).}

\item{weight.list}{weights for performing weighted subsampling sketching. Note that the weight.list is a list of weights
per batch. The weights for each batch is a vector of nonnegative values of the same size as the number of cells in the batch.
No weight by default (weight.list=NULL).}
}
\value{
a list of subsampled datasets
}
\description{
Subsample the samples in each batch, eighter weighted with the input weight.list or unweighted.
}
