% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transfer.R
\name{transfer_objective_func}
\alias{transfer_objective_func}
\title{Calculate the objective function}
\usage{
transfer_objective_func(X, W, H, lambd, b)
}
\arguments{
\item{X}{An ncells-by-ngenes gene expression matrix}

\item{W}{An ngenes-by-r reference low dimensional factor matrix}

\item{H}{A factor loading matrix of size ncells-by-r}

\item{lambd}{A numeric vecor of the scaling}

\item{b}{A numeric shift vector of size p (ngenes).}
}
\value{
numeric scalar, the value of the objective function
}
\description{
\deqn{G(H, \Lambda, b; Wref):= \|X -(H Wref^T\Lambda +  1 d^T)\|_F^2}
}
