% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transfer.R
\name{transfer_solve_lambda}
\alias{transfer_solve_lambda}
\title{Solve for dataset specific scalings lambda for target data}
\usage{
transfer_solve_lambda(X, W, H, b)
}
\arguments{
\item{X}{An ncells-by-ngenes gene expression matrix}

\item{W}{An ngenes-by-r reference low dimensional factor matrix}

\item{H}{A factor loading matrix of size ncells-by-r}

\item{b}{A numeric shift vector of size p (ngenes).}
}
\value{
numeric vector, scaling of target data with respect to the reference factor matrix
}
\description{
\deqn{argmin_{lambda>=0} ||X- HW^T diag(lambd) - 1_n b^T||_F^2}
}
