% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transfer.R
\name{transfer_solve_subproblem}
\alias{transfer_solve_subproblem}
\title{Solve subproblems via coordinate descent, given the parameter to update}
\usage{
transfer_solve_subproblem(
  params.to.update = c("lambda", "b", "H"),
  X,
  W,
  H,
  b,
  lambd,
  n.cores,
  verbose = T
)
}
\arguments{
\item{params.to.update}{a characteristic scalar, choice of (lambda','b','H'),
specifying which set of parameters to update}

\item{X}{An ncells-by-ngenes gene expression matrix}

\item{W}{An ngenes-by-r reference low dimensional factor matrix}

\item{H}{A factor loading matrix of size ncells-by-r}

\item{b}{A numeric shift vector of size p (ngenes).}

\item{verbose}{boolean scalar, whether to show extensive program logs (default TRUE)}

\item{lambda}{A numeric scaling vector of size p (ngenes)}
}
\value{
a list containing updated parameters: H, lambda,  b
}
\description{
Solve the subproblem given which parameter set to update. For each subproblem,
the exact solution is obtained
}
