% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{ari}
\alias{ari}
\title{Adjusted Rand Index}
\usage{
ari(x, y)
}
\arguments{
\item{x}{a vector containing the labels of the first classification. Must be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{y}{a vector containing the labels of the second classification.}
}
\value{
a scalar with the adjusted rand index.
}
\description{
A function to compute the adjusted rand index between two classifications.
}
