% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration.R
\name{CFITIntegrate}
\alias{CFITIntegrate}
\title{Integration of multiple data source.}
\usage{
CFITIntegrate(
  X.list,
  r = 15,
  max.niter = 100,
  tol = 1e-05,
  nrep = 1,
  init = NULL,
  future.plan = c("sequential", "transparent", "multicore", "multisession", "cluster"),
  workers = parallel::detectCores() - 1,
  verbose = T,
  seed = 0
)
}
\arguments{
\item{X.list}{a list of m ncells-by-ngenes, gene expression matrices from m data sets}

\item{r}{scalar, dimension of common factor matrix, which can be chosen as the rough number of
identifiable cells types in the joint population (default 15).}

\item{max.niter}{integer, max number of iterations (default 100).}

\item{tol}{numeric scalar, tolerance used in stopping criteria (default 1e-5).}

\item{nrep}{integer, number of repeated runs (to reduce effect of local optimum, default 1)}

\item{init}{a list of parameters for parameter initialization. The list either contains all
parameter sets: W,lambda.list, b.list, H.list, or only W will be used if provided (default NULL).}

\item{future.plan}{plan for future parallel computation, can be chosen from 'sequential','transparent','multicore','multisession' and 'cluster'. Default is 'sequential'. Note that Rstudio does not support 'multicore'.}

\item{workers}{additional parameter for \code{future::plan()}, in cases of 'multicore','multisession' and 'cluster'.}

\item{verbose}{boolean scalar, whether to show extensive program logs (default TRUE)}

\item{seed}{random seed used (default 0)}
}
\value{
a list containing  \describe{
\item{W}{ngenes-by-r numeric matrix, estimated common factor matrix}
\item{H.list}{A list of m factor loading matrix of size ncells-by-r, estimated factor loading matrices}
\item{b.list}{A list of estimated shift vector of size p (ngenes).}
\item{lambda.list}{A list of estimated scaling vector of size p (ngenes).}
\item{convergence}{boolean, whether the algorithm converge}
\item{obj}{numeric scalar, value of the objective function at convergence or when maximum iteration achieved}
\item{niter}{integer, the iteration at convergence (or maximum iteration if not converge)}
\item{delta}{numeric scalar, the relative difference in last objective update}
\item{params}{list of parameters used for the algorithm: max.iter, tol, nrep}
}
}
\description{
Solve the model parameters through Iterative Nonnegative Matrix Factorization (iNMF),
by minimizing the objective function \deqn{1/N \sum_j||X_j -(H_jW^T\Lambda_j  + 1_{n_j} b_j^T)||_F^2} with penalties.
}
